three backend files merged into one - index.js
