+++ backend/server.js
@@ -1,21 +1,23 @@
-import express, { json } from 'express'
-import dotenv from 'dotenv'
-dotenv.config({ path: './.env' })
-import auth from './routes/auth.js'
-import cors from'cors'
+// import express, { json } from 'express'
+// import dotenv from 'dotenv'
+// dotenv.config({ path: './.env' })
+// import auth from './routes/auth.js'
+// import cors from'cors'
+// import fetch from 'node-fetch'
+// import auth 
 
-// Load environment variables from the .env file in the backend directory
+// // Load environment variables from the .env file in the backend directory
 
-const app = express()
-const port = '3002'
+// const app = express()
+// const port = '3002'
 
-app.use(cors());
-app.use(express.json());
+// app.use(cors());
+// app.use(express.json());
 
-app.get('/', (req, res) => {
-    res.send("Hello world")
-});
-app.use('/api', auth)
-app.listen(port, () => {
-    console.log('Server running on port 3002');
-});
+// app.get('/', (req, res) => {
+//     res.send("Hello world")
+// });
+// app.use('/api', auth)
+// app.listen(port, () => {
+//     console.log('Server running on port 3002');
+// });
